small fix for pretty-printing & minor changes
git-svn-id: http://k-framework.googlecode.com/svn/trunk@7356 82a86efa-2583-11df-8ea0-49ba94de29ab

Former-commit-id: d2cc8dce12368337973d9ca77cf62a240f64ef79 [formerly afe48bd0f5b8956b7e6f4db73ce4ea6964702083]
Former-commit-id: 39b89b21dd1305b8a910c1d1b486b2a37491a3f9

+++ javasources/JKRun/src/ro/uaic/info/fmse/jkrun/FileUtil.java
@@ -106,11 +106,13 @@ public class FileUtil {
 	public static boolean deleteDirectory(File path) {
 		if (path.exists()) {
 			File[] files = path.listFiles();
-			for (int i = 0; i < files.length; i++) {
-				if (files[i].isDirectory()) {
-					deleteDirectory(files[i]);
-				} else {
-					files[i].delete();
+			if (files != null && files.length > 0) {
+				for (int i = 0; i < files.length; i++) {
+					if (files[i].isDirectory()) {
+						deleteDirectory(files[i]);
+					} else {
+						files[i].delete();
+					}
 				}
 			}
 		}
@@ -359,6 +361,12 @@ public class FileUtil {
 		while (index != -1) {
 			index = sb.indexOf("_", index);
 			if (index != -1) {
+				//when the number of underscores is greater than the number of children (such in the case of family.logik)
+				if (count >= elements.size()) {		
+					sb.insert(index, "");
+					sb = sb.deleteCharAt(index);
+					break;		
+				}
 				String s = (String)elements.get(count);
 				sb.insert(index, s);
 				index += s.length();
